Update .gitignore to ignore dist-folder
Trying to fix https://unpkg.com/svelthree not being usable (importable from) in a svelte.dev REPL environment. One thing I didn't do (although the unpkg instructions state it should be done) is ignoring the dist-folder. So maybe this fixes it somehow? Not sure though why it should, because the unpkg link is linking the correct version.